+++ src/services/queries/upload-revenue/useRevenueUploadHistory.ts
@@ -2,39 +2,9 @@ import {
   QueryClient, UseQueryResult, useMutation, useQuery,
 } from "@tanstack/react-query";
 
-import { MOCK_TRANSACTION_UPLOAD } from "@/constants/mock";
+import { IGETTransactionUploadResponse } from "@/services/api/types/transaction";
+import { deleteRevenueUploadHistory, getRevenueUploadHistory, postRevenueUploadHistory } from "@/services/api/upload-revenue-mock-api";
 import { MEMBER_TYPE } from "@/types/enums/user.enum";
-
-import { IGETTransactionUploadResponse } from "../../api/types/transaction";
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-const getRevenueUploadHistory = (month: string): Promise<IGETTransactionUploadResponse> => {
-  // TODO: (GET) 정산 업로드 내역 가져오기
-  return new Promise((resolve) => {
-    setTimeout(() => {
-      resolve(MOCK_TRANSACTION_UPLOAD);
-    }, 3000);
-  });
-};
-
-const deleteRevenueUploadHistory = () => {
-  // TODO: (DELETE) 정산 업로드 내역 삭제
-  return new Promise((resolve) => {
-    setTimeout(() => {
-      resolve("업로드 내역이 삭제되었습니다.");
-    }, 3000);
-  });
-};
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-const postRevenueUploadHistory = (file: File, uploadAt: string) => {
-  // TODO: (POST) 정산 내역 업로드
-  return new Promise((resolve) => {
-    setTimeout(() => {
-      resolve("정산 내역 업로드가 완료되었습니다.");
-    }, 3000);
-  });
-};
 
 /* 정산 업로드 내역 GET */
 const useUploadHistoryGet = (month: string) => {
@@ -76,8 +46,6 @@ interface FileData {
 }
 
 const useUploadHistoryPost = (
-  // file: File,
-  // uploadAt: string,
   queryClient: QueryClient,
   showToast: (message: string) => void,
 ) => {
@@ -95,6 +63,4 @@ const useUploadHistoryPost = (
   return { postUploadHistory, isLoading };
 };
 
-export {
-  getRevenueUploadHistory, useUploadHistoryGet, useUploadHistoryDelete, useUploadHistoryPost,
-};
+export { useUploadHistoryGet, useUploadHistoryDelete, useUploadHistoryPost };
